pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "customer-engine"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
authors = [{ name = "Author name", email = "author@email.com" }]
dependencies = [
    "lego-workflows < 0.1.0",
    "sqlalchemy-libsql",
    "python-dotenv < 2",
    "fastapi < 1",
    "uvicorn[standard] < 1",
    "httpx < 1",
    "gunicorn < 22",
    "mashumaro < 4",
]

[project.urls]
Documentation = "https://github.com/pymtech-co/whatsapp_customer_engine#readme"
Issues = "https://github.com/pymtech-co/whatsapp_customer_engine/issues"
Source = "https://github.com/pymtech-co/whatsapp_customer_engine"


[project.optional-dependencies]
dev = ["pyrgo < 3", "pytest-asyncio", "types-toml"]

[project.scripts]
custeng = "customer_engine.cli:cli"

[tool.ruff]
line-length = 88
fix = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
ignore = [
    "ANN101",
    "D203",
    "D212",
    "COM812",
    "COM819",
    "D206",
    "E501",
    "ISC001",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
]
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["INP001", "S101", "D"]
"scripts/*.py" = ["INP001"]

[tool.ruff.lint.isort]
known-first-party = ["customer_engine"]
combine-as-imports = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = ["--import-mode=importlib", "--strict-markers"]
markers = ["integration: mark integration tests.", "unit: mark unittest."]
asyncio_mode = "auto"

[tool.mypy]
follow_imports = "normal"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true

[tool.pyrgo]
extra-paths = ["scripts"]
